requirements.txt
openai
langchain_nvidia_ai_endpoints
langchain_community
faiss-cpu
python-dotenv
streamlit
pypdf
#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